Extract lex to file
